+++ src/demo-app/features/modal-search/modal-search-demo.ts
@@ -1,0 +1,12 @@
+import {Component} from '@angular/core';
+
+@Component({
+  moduleId: module.id,
+  providers: [],
+  templateUrl: './modal-search-demo.html',
+})
+export class ModalSearchDemo {
+
+
+}
+