+++ src/lib/index.ts
@@ -7,3 +7,4 @@ export * from './spinner/index';
 export * from './dropdown/index';
 export * from './option/index';
 export * from './progress-bar/index';
+export * from './modal-search/index';